Decompose <App /> component for easier testing
