+++ src/components/MainPage.js
@@ -1,0 +1,33 @@
+import React, { Component } from 'react';
+// import { setSearchField, requestRobots } from '../actions';
+
+import CardList from '../components/CardList';
+import SearchBox from '../components/SearchBox';
+import Scroll from '../components/Scroll';
+import ErrorBoundry from '../components/ErrorBoundry';
+import Header from '../components/Header';
+
+import './MainPage.css';
+
+class MainPage extends Component {
+  render() {
+    const { robots, searchField, onSearchChange, isPending } = this.props;
+    const filteredRobots = robots.filter(robot => {
+      return robot.name.toLowerCase().includes(searchField.toLowerCase());
+    })
+    return (
+      <div className='tc'>
+        <SearchBox searchChange={onSearchChange} />
+        <Scroll>
+          {isPending ? <h1>Loading</h1> :
+            <ErrorBoundry>
+              <CardList robots={filteredRobots} />
+            </ErrorBoundry>
+          }
+        </Scroll>
+      </div>
+    );
+  }
+}
+
+export default MainPage;
